correzione bug grafico in profile activity e correzione bug in scaricamento dato utente "altezza raggiunta"

+++ src/org/unipd/nbeghin/climbtheworld/ClimbApplication.java
@@ -1127,7 +1127,7 @@ public class ClimbApplication extends Application{
 									// Hooray! The user is logged in.
 									me.setLevel(user.getInt("level"));
 			    		    				me.setXP(user.getInt("XP"));
-			    		    				me.setHeight(user.getLong("height"));
+			    		    				me.setHeight(user.getDouble("height"));
 			    		    				JSONObject stats = user.getJSONObject("mean_daily_steps");
 			    		    				if (stats != null && stats.length() > 0) {
 			    		    				try {
